pyproject.toml
[project]
name = "glamira-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.11,<3.14"
dependencies = [
    "pymongo (>=4.16.0,<5.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "ip2location (>=8.11.0,<9.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "playwright (>=1.58.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
